% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{figure_1_df}
\alias{figure_1_df}
\title{Data frame used to construct Figure 1}
\format{
A data frame with 5 columns and 12877 rows
}
\usage{
figure_1_df
}
\description{
Figure 1 is the baseline mortality map.
}
\keyword{datasets}
