% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc}
\alias{misc}
\alias{trunc_dec}
\alias{lm_eqn}
\alias{std_err}
\alias{mod_10_time}
\alias{get_McNellis_data}
\title{Miscellaneous small functions}
\usage{
trunc_dec(x, type, level = 1)

lm_eqn(df)

std_err(x)

mod_10_time(y)

get_McNellis_data()
}
\description{
Small functions for McNellis2021.
}
