% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_functions}
\alias{plot_functions}
\alias{Multiplot}
\alias{PlotOverallImportance}
\alias{PlotAgentImportance}
\alias{PlotFeatureContrib}
\alias{PlotPartialDepend}
\alias{PlotCorr}
\alias{HighResTiff}
\alias{plot_diff_mort_map}
\alias{make_mort_bar_plot}
\title{Functions to create (mostly) final plots for McNellis2021 manuscript.}
\usage{
Multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)

PlotOverallImportance(data_df)

PlotAgentImportance(data_df)

PlotFeatureContrib(
  plot_df,
  ylab0,
  panel_lab = NULL,
  panel_coord = NULL,
  xlim0 = NULL,
  ylim0 = NULL,
  order_descending = TRUE
)

PlotPartialDepend(plot_df, y0, order_descending = T)

PlotCorr(plot_df)

HighResTiff(plot_obj, file, width_in, height_in, resolution_dpi)

plot_diff_mort_map(
  mort_df,
  diff_var = "mort_rate",
  legend_title = NA,
  scale_adjust = 1
)

make_mort_bar_plot(plot_df)
}
\description{
Input data comes from the `McNellis2021/data`` directory.

Setup script is in McNellis2021/exec/make_figures.R

See the README for additional details.
}
\author{
Brandon McNellis
}
